requirement.txt
appdirs==1.4.4
asgiref==3.2.7
auxlib==0.0.43
cffi==1.14.3
conda==4.2.7
cryptography==3.1.1
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.5.0
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.1
filelock==3.0.12
gunicorn==20.0.4
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
toolz==0.10.0
urllib3==1.25.10
virtualenv==20.0.20
whitenoise==5.2.0
